requirements.txt
# ── LLM Inference ─────────────────────────────────────────────────────────
# vLLM for primary inference (continuous batching, true parallelism)
vllm>=0.4.0
# OpenAI-compatible client (works with vLLM's OpenAI API server)
openai>=1.30.0
# Async HTTP (parallel requests to vLLM)
httpx>=0.27.0
# Ollama for embeddings only
ollama>=0.3.0

# ── Document Parsing ──────────────────────────────────────────────────────
PyMuPDF>=1.24.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# ── Text Processing ───────────────────────────────────────────────────────
regex>=2024.0.0
tiktoken>=0.7.0

# ── Vector Store ──────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Agent Orchestration ───────────────────────────────────────────────────
langgraph>=0.2.0
langchain-core>=0.3.0

# ── Utilities ─────────────────────────────────────────────────────────────
rich>=13.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.12.0
